Move GitHub parsing to encodingconverter
